Delete EmbeddedFileManager module
This module job was to store manifests as constants inside memory but it
is not necessary as the constants content were then written to files
again before applying by kubectl. So the only thing it in fact did was
creating middleware layer that read files from one directory and save
them later in different, hence not giving any value to codebase but
rather providing code to maintain.

Signed-off-by: Rafal Lal <rafal.lal@tietoevry.com>