Add 2 workers in all the clusters
It also waits 5m the cluster to be ready to avoid
race conditions and updates kind to its last release.

It shows all compile warnings and debug data to allow
us to debug our gate.

Please not it doesn't add workers when testing private
registries as the spec tests must be rewritten first.

Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>